+++ tests/api-test-test.js
@@ -1,11 +1,14 @@
 'use strict';
 
-const { API } = require('@janiscommerce/api');
+const { API, Handler } = require('@janiscommerce/api');
+const { ApiSession } = require('@janiscommerce/api-session');
 
 const assert = require('assert');
 
 const { APITest, APITestError } = require('../lib');
 const APITestCaller = require('..');
+
+const makeHandler = api => (...args) => new Handler(api).handle(...args);
 
 describe('APITest', async () => {
 
@@ -22,8 +25,8 @@ describe('APITest', async () => {
 
 		async process() {
 			this
-				.setHeaders({ 'some-header': 123 })
-				.setCookies({ 'some-cookie': 321 })
+				.setHeaders({ 'some-header': '123' })
+				.setCookies({ 'some-cookie': '321' })
 				.setBody({ a: 2 })
 				.setCode(200);
 		}
@@ -43,18 +46,30 @@ describe('APITest', async () => {
 		}
 	}
 
-	class APIWithErrorAndCustomBody extends API {
+	class APIClientClass extends API {
 
-		async validate() {
-			const error = new Error('Some error');
-			error.body = { custom: true };
-			throw error;
+		async process() {
+
+			const client = await this.session.client;
+			delete client.getInstance;
+
+			this
+				.setBody({
+					clientCode: this.session.clientCode,
+					client
+				})
+				.setCode(200);
 		}
 	}
 
-	APITestCaller(APIClass, '/custom/path/to/my/api', [{
+	const APIHandler = makeHandler(APIClass);
+	const APICompleteHandler = makeHandler(APICompleteClass);
+	const APIWithError5xxHandler = makeHandler(APIWithError5xx);
+	const APIWithError4xxHandler = makeHandler(APIWithError4xx);
+	const APIClientHandler = makeHandler(APIClientClass);
+
+	APITestCaller(APIHandler, '/custom/path/to/my/api', [{
 		description: 'Should response an empty body and a 200 http code using a endpoint',
-		prepare: () => {},
 		request: {},
 		response: {
 			code: 200,
@@ -74,7 +89,8 @@ describe('APITest', async () => {
 		it('when rules given aren\'t an array', () => {
 
 			notAnArray.forEach(rules => {
-				assert.throws(() => APITestCaller(APIClass, rules), {
+				assert.throws(() => new APITest(APIHandler, rules), {
+					message: 'Rules error - invalid format, must be an array',
 					name: 'APITestError',
 					code: APITestError.codes.INVALID_RULES
 				});
@@ -83,8 +99,10 @@ describe('APITest', async () => {
 
 		it('when rule is not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(rule => {
-				assert.throws(() => APITestCaller(APIClass, [rule]), {
+				assert.throws(() => apiTest.validateRule(rule), {
+					message: 'Rule error - must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_FORMAT
 				});
@@ -93,7 +111,9 @@ describe('APITest', async () => {
 
 		it('when rule description is missing', () => {
 
-			assert.throws(() => APITestCaller(APIClass, [{}]), {
+			const apiTest = new APITest(APIHandler);
+			assert.throws(() => apiTest.validateRule({}), {
+				message: 'Rule error - invalid description, must be a string',
 				name: 'APITestError',
 				code: APITestError.codes.RULE_INVALID_DESCRIPTION
 			});
@@ -101,8 +121,10 @@ describe('APITest', async () => {
 
 		it('when rule description is not a string', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAString.forEach(description => {
-				assert.throws(() => APITestCaller(APIClass, [{ description }]), {
+				assert.throws(() => apiTest.validateRule({ description }), {
+					message: 'Rule error - invalid description, must be a string',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_DESCRIPTION
 				});
@@ -111,28 +133,23 @@ describe('APITest', async () => {
 
 		it('when rule session is not an object nor a boolean', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObjectNorBoolean.forEach(session => {
-				assert.throws(() => APITestCaller(APIClass, [{ description: 'foo', session }]), {
+				assert.throws(() => apiTest.validateRule({ description: 'foo', session }), {
+					message: 'Rule error - invalid session, must be an object or a boolean',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_SESSION
 				});
 			});
 		});
 
-		it('when client request is not an object', () => {
-
-			notAnObject.forEach(async client => {
-				assert.throws(() => APITestCaller(APIClass, [{ description: 'foo', client }]), {
-					name: 'APITestError',
-					code: APITestError.codes.RULE_INVALID_CLIENT
-				});
-			});
-		});
 
 		it('when rule request is not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(request => {
-				assert.throws(() => APITestCaller(APIClass, [{ description: 'foo', request }]), {
+				assert.throws(() => apiTest.validateRule({ description: 'foo', request }), {
+					message: 'Rule error - invalid request, must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_REQUEST
 				});
@@ -141,8 +158,10 @@ describe('APITest', async () => {
 
 		it('when rule request.rawData is not aa string', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAString.forEach(rawData => {
-				assert.throws(() => APITestCaller(APIClass, [{ description: 'foo', request: { rawData } }]), {
+				assert.throws(() => apiTest.validateRule({ description: 'foo', request: { rawData } }), {
+					message: 'Rule error - invalid request rawData, must be a string',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_REQUEST
 				});
@@ -150,7 +169,10 @@ describe('APITest', async () => {
 		});
 
 		it('when rule response is missing', () => {
-			assert.throws(() => APITestCaller(APIClass, [{ description: 'foo' }]), {
+
+			const apiTest = new APITest(APIHandler);
+			assert.throws(() => apiTest.validateRule({ description: 'foo' }), {
+				message: 'Rule error - invalid response, must be an object',
 				name: 'APITestError',
 				code: APITestError.codes.RULE_INVALID_RESPONSE
 			});
@@ -158,8 +180,10 @@ describe('APITest', async () => {
 
 		it('when rule response is not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(response => {
-				assert.throws(() => APITestCaller(APIClass, [{ description: 'foo', response }]), {
+				assert.throws(() => apiTest.validateRule({ description: 'foo', response }), {
+					message: 'Rule error - invalid response, must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_RESPONSE
 				});
@@ -168,14 +192,14 @@ describe('APITest', async () => {
 
 		it('when rule response code is not a number', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notANumber.forEach(code => {
 
-				const rules = [{
+				assert.throws(() => apiTest.validateRule({
 					description: 'foo',
 					response: { code }
-				}];
-
-				assert.throws(() => APITestCaller(APIClass, rules), {
+				}), {
+					message: 'Rule error - invalid response code, must be a number',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_RESPONSE_CODE
 				});
@@ -184,14 +208,14 @@ describe('APITest', async () => {
 
 		it('when rule response headers are not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(headers => {
 
-				const rules = [{
+				assert.throws(() => apiTest.validateRule({
 					description: 'foo',
 					response: { headers }
-				}];
-
-				assert.throws(() => APITestCaller(APIClass, rules), {
+				}), {
+					message: 'Rule error - invalid response headers, must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_RESPONSE_HEADERS
 				});
@@ -200,14 +224,14 @@ describe('APITest', async () => {
 
 		it('when rule response strict headers are not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(strictHeaders => {
 
-				const rules = [{
+				assert.throws(() => apiTest.validateRule({
 					description: 'foo',
 					response: { strictHeaders }
-				}];
-
-				assert.throws(() => APITestCaller(APIClass, rules), {
+				}), {
+					message: 'Rule error - invalid response strict headers, must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_RESPONSE_HEADERS
 				});
@@ -216,14 +240,14 @@ describe('APITest', async () => {
 
 		it('when rule response cookies are not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(cookies => {
 
-				const rules = [{
+				assert.throws(() => apiTest.validateRule({
 					description: 'foo',
 					response: { cookies }
-				}];
-
-				assert.throws(() => APITestCaller(APIClass, rules), {
+				}), {
+					message: 'Rule error - invalid response cookies, must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_RESPONSE_COOKIES
 				});
@@ -232,14 +256,14 @@ describe('APITest', async () => {
 
 		it('when rule response strict cookies are not an object', () => {
 
+			const apiTest = new APITest(APIHandler);
 			notAnObject.forEach(strictCookies => {
 
-				const rules = [{
+				assert.throws(() => apiTest.validateRule({
 					description: 'foo',
 					response: { strictCookies }
-				}];
-
-				assert.throws(() => APITestCaller(APIClass, rules), {
+				}), {
+					message: 'Rule error - invalid response strict cookies, must be an object',
 					name: 'APITestError',
 					code: APITestError.codes.RULE_INVALID_RESPONSE_COOKIES
 				});
@@ -247,9 +271,8 @@ describe('APITest', async () => {
 		});
 	});
 
-	APITestCaller(APIClass, [{
+	APITestCaller(APIHandler, [{
 		description: 'Should response an empty body and a 200 http code',
-		prepare: () => {},
 		request: {},
 		response: {
 			code: 200,
@@ -257,9 +280,8 @@ describe('APITest', async () => {
 		}
 	}]);
 
-	APITestCaller(APIWithError5xx, [{
+	APITestCaller(APIWithError5xxHandler, [{
 		description: 'Should response body with message and a 500 http code',
-		prepare: () => {},
 		request: {},
 		response: {
 			code: 500,
@@ -267,9 +289,8 @@ describe('APITest', async () => {
 		}
 	}]);
 
-	APITestCaller(APIWithError4xx, [{
+	APITestCaller(APIWithError4xxHandler, [{
 		description: 'Should response body with message and a 400 http code',
-		prepare: () => {},
 		request: {},
 		response: {
 			code: 400,
@@ -277,19 +298,9 @@ describe('APITest', async () => {
 		}
 	}]);
 
-	APITestCaller(APIWithErrorAndCustomBody, [{
-		description: 'Should response body with message and a 400 http code',
-		prepare: () => {},
-		request: {},
-		response: {
-			code: 400,
-			body: { custom: true }
-		}
-	}]);
+	it('Should reject when response code doesn\'t match', async () => {
 
-	it('Should reject when response codes doesn\'t match', async () => {
-
-		const apiTest = new APITest(APIClass);
+		const apiTest = new APITest(APIHandler);
 
 		await assert.rejects(() => apiTest.assert({
 			response: { code: 999 }
@@ -308,7 +319,8 @@ describe('APITest', async () => {
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
 			response: { code: 200, body: 10 }
@@ -323,14 +335,15 @@ describe('APITest', async () => {
 
 		class MyAPIClass extends API {
 			async process() {
-				this.setHeader('my-header', 123);
+				this.setHeader('my-header', '123');
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
-			response: { code: 200, headers: { 'other-header': 123 } }
+			response: { code: 200, headers: { 'other-header': '123' } }
 		}), {
 			// the assert is assert(typeof response.headers[name] !== 'undefined');
 			code: 'ERR_ASSERTION',
@@ -343,19 +356,20 @@ describe('APITest', async () => {
 
 		class MyAPIClass extends API {
 			async process() {
-				this.setHeaders({ 'my-header': 123, 'other-header': 321 });
+				this.setHeaders({ 'my-header': '123', 'other-header': '321' });
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
-			response: { code: 200, headers: { 'my-header': 321 } }
+			response: { code: 200, headers: { 'my-header': '321' } }
 		}), {
 			// header found but with other value
 			code: 'ERR_ASSERTION',
-			actual: 123,
-			expected: 321
+			actual: '123',
+			expected: '321'
 		});
 	});
 
@@ -363,19 +377,20 @@ describe('APITest', async () => {
 
 		class MyAPIClass extends API {
 			async process() {
-				this.setHeaders({ 'my-header': 123, 'other-header': 321 });
+				this.setHeaders({ 'my-header': '123', 'other-header': '321' });
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
-			response: { code: 200, strictHeaders: { 'my-header': 123 } }
+			response: { code: 200, strictHeaders: { 'my-header': '123' } }
 		}), {
 			// header found but with other value
 			code: 'ERR_ASSERTION',
-			actual: { 'my-header': 123, 'other-header': 321 },
-			expected: { 'my-header': 123 }
+			actual: { 'my-header': '123', 'other-header': '321' },
+			expected: { 'my-header': '123' }
 		});
 	});
 
@@ -383,14 +398,15 @@ describe('APITest', async () => {
 
 		class MyAPIClass extends API {
 			async process() {
-				this.setCookie('my-cookie', 123);
+				this.setCookie('my-cookie', '123');
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
-			response: { code: 200, cookies: { 'other-cookie': 123 } }
+			response: { code: 200, cookies: { 'other-cookie': '123' } }
 		}), {
 			// the assert is assert(typeof response.cookies[name] !== 'undefined');
 			code: 'ERR_ASSERTION',
@@ -403,19 +419,20 @@ describe('APITest', async () => {
 
 		class MyAPIClass extends API {
 			async process() {
-				this.setCookies({ 'my-cookie': 123, 'other-cookie': 321 });
+				this.setCookies({ 'my-cookie': '123', 'other-cookie': '321' });
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
-			response: { code: 200, cookies: { 'my-cookie': 321 } }
+			response: { code: 200, cookies: { 'my-cookie': '321' } }
 		}), {
 			// cookie found but with other value
 			code: 'ERR_ASSERTION',
-			actual: 123,
-			expected: 321
+			actual: '123',
+			expected: '321'
 		});
 	});
 
@@ -423,32 +440,33 @@ describe('APITest', async () => {
 
 		class MyAPIClass extends API {
 			async process() {
-				this.setCookies({ 'my-cookie': 123, 'other-cookie': 321 });
+				this.setCookies({ 'my-cookie': '123' });
 			}
 		}
 
-		const apiTest = new APITest(MyAPIClass);
+		const MyAPIHandler = makeHandler(MyAPIClass);
+		const apiTest = new APITest(MyAPIHandler);
 
 		await assert.rejects(() => apiTest.assert({
-			response: { code: 200, strictCookies: { 'my-cookie': 123 } }
+			response: { code: 200, strictCookies: {} }
 		}), {
 			// cookie found but with other value
 			code: 'ERR_ASSERTION',
-			actual: { 'my-cookie': 123, 'other-cookie': 321 },
-			expected: { 'my-cookie': 123 }
+			actual: { 'my-cookie': '123' },
+			expected: {}
 		});
 	});
 
-	APITestCaller(APICompleteClass, [{
+	APITestCaller(APICompleteHandler, [{
 		description: 'Should set response code, body, headers and cookie',
 		before: () => {},
 		after: () => {},
 		request: {
 			endpoint: 'custom-endpoint',
 			data: { fooData: 1 },
-			pathParameters: [1, 2],
-			headers: { 'some-header': 123 },
-			cookies: { 'some-cookie': 321 }
+			pathParameters: ['1', '2'],
+			headers: { 'some-header': '123' },
+			cookies: { 'some-cookie': '321' }
 		},
 		getResponse: () => {
 			// do something great with the response received as a param
@@ -456,30 +474,17 @@ describe('APITest', async () => {
 		response: {
 			code: 200,
 			body: { a: 2 },
-			strictHeaders: { 'some-header': 123 },
-			strictCookies: { 'some-cookie': 321 }
+			strictHeaders: { 'some-header': '123' },
+			strictCookies: { 'some-cookie': '321' }
 		}
 	}]);
 
-	class APIClientClass extends API {
-
-		async process() {
-
-			const client = await this.session.client;
-			delete client.getInstance;
-
-			this
-				.setBody({
-					clientCode: this.session.clientCode,
-					client
-				})
-				.setCode(200);
-		}
-	}
-
-	APITestCaller(APIClientClass, [{
+	APITestCaller(APIClientHandler, [{
 		description: 'Should set response code and body with default session\'s client',
 		session: true,
+		before: /** @param {import('sinon').SinonStatic} sinon */ sinon => {
+			sinon.stub(ApiSession.prototype, 'client').value(Promise.resolve({ code: 'defaultClient', id: '507f191e810c19729de860ea' }));
+		},
 		request: {
 			endpoint: 'custom-endpoint'
 		},
@@ -489,17 +494,20 @@ describe('APITest', async () => {
 				clientCode: 'defaultClient',
 				client: {
 					code: 'defaultClient',
-					id: 1
+					id: '507f191e810c19729de860ea'
 				}
 			}
 		}
 	}]);
 
-	APITestCaller(APIClientClass, [{
+	APITestCaller(APIClientHandler, [{
 		description: 'Should set response code, body with current injected client code',
 		session: {
-			clientId: 1,
+			clientId: '507f191e810c19729de860eb',
 			clientCode: 'my-client-code'
+		},
+		before: /** @param {import('sinon').SinonStatic} sinon */ sinon => {
+			sinon.stub(ApiSession.prototype, 'client').value(Promise.resolve({ code: 'my-client-code', id: '507f191e810c19729de860eb' }));
 		},
 		request: {
 			endpoint: 'custom-endpoint'
@@ -509,35 +517,8 @@ describe('APITest', async () => {
 			body: {
 				clientCode: 'my-client-code',
 				client: {
-					code: 'defaultClient',
-					id: 1
-				}
-			}
-		}
-	}]);
-
-	APITestCaller(APIClientClass, [{
-		description: 'Should set response code, body with current injected client code and client',
-		session: {
-			clientId: 1,
-			clientCode: 'my-client-code'
-		},
-		client: {
-			id: 1,
-			code: 'my-client-code',
-			customField: true
-		},
-		request: {
-			endpoint: 'custom-endpoint'
-		},
-		response: {
-			code: 200,
-			body: {
-				clientCode: 'my-client-code',
-				client: {
-					id: 1,
 					code: 'my-client-code',
-					customField: true
+					id: '507f191e810c19729de860eb'
 				}
 			}
 		}