+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Default session client, user and profile IDs are now ObjectIds instead of integers **BREAKING CHANGE**
+- Rule `client` is no longer supported. You must stub API session yourself. For example `sinon.stub(ApiSession.prototype, 'client').value(Promise.resolve({ /* yourClientData */ }));`.
 
 ## [6.0.0] - 2023-09-18
 ### Changed